rename `with-serde` feature to `serde`
