+++ rrule/Cargo.toml
@@ -20,11 +20,11 @@ log = "0.4.16"
 regex = "1.5.5"
 clap = { version = "3.1.15", optional = true, features = ["derive"] }
 thiserror = "1.0.30"
-serde = { version = "1.0.137", optional = true, features = ["derive"] }
-serde_with = { version = "1.13.0", optional = true, features = ["macros"] }
+serde_with = { version = "1.13.0", optional = true }
 
 [dev-dependencies]
 serde_json = "1.0.80"
+orig_serde = { package = "serde", version = "1.0.137", default-features = false }
 
 [[bin]]
 name = "rrule"
@@ -50,4 +50,4 @@ by-easter = []
 cli-tool = ["clap"]
 
 # Enable serde for some of the public structs
-with-serde = ["serde", "serde_with", "chrono/serde", "chrono-tz/serde"]
+serde = ["serde_with", "chrono/serde", "chrono-tz/serde"]
