+++ rrule/tests/rrule.rs
@@ -4,7 +4,6 @@ use chrono::{Datelike, TimeZone};
 use chrono_tz::UTC;
 use common::{test_recurring_rrule, ymd_hms};
 use rrule::{DateFilter, Frequency, NWeekday, RRule, RRuleProperties, Weekday};
-use std::str::FromStr;
 
 #[test]
 fn yearly() {
@@ -4795,47 +4794,4 @@ fn test_between_inclusive_both_hit() {
         vec![before, middle, after],
         rrule.all_between(before, after, inc).unwrap()
     );
-}
-
-#[test]
-fn rrule_to_and_from_str() {
-    let test_cases = [
-        "DTSTART:20120201T093000Z\nRRULE:FREQ=YEARLY;COUNT=3",
-        "DTSTART:20120201T093000Z\nRRULE:FREQ=WEEKLY;INTERVAL=5;BYDAY=-2MO,FR",
-        "DTSTART;TZID=America/New_York:19000201T093000Z\nRRULE:UNTIL=19990404T110000Z;FREQ=WEEKLY;BYDAY=TU,WE",
-    ];
-
-    for test_str in test_cases {
-        let test_obj = RRule::from_str(test_str).unwrap();
-        let test_str = test_obj.to_string();
-        let test_obj2 = RRule::from_str(&test_str).unwrap();
-
-        assert_eq!(test_obj, test_obj2);
-    }
-}
-
-#[cfg(feature = "with-serde")]
-#[test]
-fn serialize_deserialize_json_to_and_from_rrule() {
-    #[allow(dead_code)]
-    #[derive(serde::Deserialize, serde::Serialize, PartialEq, Debug)]
-    struct RruleTest {
-        rrule: RRule,
-    }
-
-    let test_cases = [
-        "DTSTART:20120201T093000Z\nRRULE:FREQ=DAILY;COUNT=5",
-        "DTSTART:20120201T093000Z\nRRULE:FREQ=WEEKLY;INTERVAL=5;BYDAY=-2MO,FR",
-        "DTSTART;TZID=America/New_York:19000201T093000Z\nRRULE:UNTIL=19990404T110000Z;FREQ=WEEKLY;BYDAY=TU,WE",
-    ];
-
-    for test_str in test_cases {
-        let rrule = RRule::from_str(test_str).unwrap();
-        let src_obj = RruleTest { rrule };
-
-        let test_str = serde_json::to_string(&src_obj).unwrap();
-        let final_obj = serde_json::from_str::<RruleTest>(&test_str).unwrap();
-
-        assert_eq!(src_obj, final_obj);
-    }
 }
